app/requirements.txt
fastapi
uvicorn
jinja2
passlib
python-multipart
pyjwt
sqlalchemy